refactor: Delete unused file
