--- feature/auth/src/main/java/com/teamoffroad/feature/auth/domain/model/UserGender.kt
@@ -1,5 +1,0 @@
-package com.teamoffroad.feature.auth.domain.model
-
-enum class UserGender {
-    MALE, FEMALE, OTHER,
-}
