+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/component/BirthDateHintText.kt
@@ -6,7 +6,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import com.teamoffroad.core.designsystem.theme.ErrorNew
 import com.teamoffroad.core.designsystem.theme.OffroadTheme
-import com.teamoffroad.feature.auth.presentation.model.SetBirthDateUiState
 
 @Composable
 fun BirthDateHintText(
