--- feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/model/SetBirthDateUiState.kt
@@ -1,11 +1,0 @@
-package com.teamoffroad.feature.auth.presentation.model
-
-data class SetBirthDateUiState(
-    val year: String = "",
-    val month: String = "",
-    val day: String = "",
-    val date: String = "",
-    val yearValidateResult: DateValidateResult = DateValidateResult.Empty,
-    val monthValidateResult: DateValidateResult = DateValidateResult.Empty,
-    val dayValidateResult: DateValidateResult = DateValidateResult.Empty,
-)
