--- feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/model/SetNicknameUiState.kt
@@ -1,6 +1,0 @@
-package com.teamoffroad.feature.auth.presentation.model
-
-data class SetNicknameUiState(
-    val nickname: String = "",
-    val nicknameValidateResult: NicknameValidateResult = NicknameValidateResult.Empty,
-)
