+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/signin/SignInScreen.kt
@@ -36,7 +36,6 @@ import com.teamoffroad.core.designsystem.theme.Main1
 import com.teamoffroad.core.designsystem.theme.Main2
 import com.teamoffroad.core.designsystem.theme.OffroadTheme
 import com.teamoffroad.core.designsystem.theme.White
-import com.teamoffroad.feature.auth.presentation.OAuthEntryPoint
 import com.teamoffroad.offroad.feature.auth.R
 import dagger.hilt.android.EntryPointAccessors
 import kotlinx.coroutines.flow.collectLatest
