+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/presentation/signup/SetCharacterViewModel.kt
@@ -1,6 +1,5 @@
-package com.teamoffroad.feature.auth.presentation
+package com.teamoffroad.feature.auth.presentation.signup
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.teamoffroad.feature.auth.domain.model.Character
